+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/components/svg/TextView.java
@@ -31,14 +31,14 @@ import static versioned.host.exp.exponent.modules.api.components.svg.TextPropert
 @SuppressLint("ViewConstructor")
 class TextView extends GroupView {
     SVGLength mTextLength = null;
-    String mBaselineShift = null;
+    private String mBaselineShift = null;
     TextLengthAdjust mLengthAdjust = TextLengthAdjust.spacing;
-    AlignmentBaseline mAlignmentBaseline;
-    @Nullable ArrayList<SVGLength> mPositionX;
-    @Nullable ArrayList<SVGLength> mPositionY;
-    @Nullable ArrayList<SVGLength> mRotate;
-    @Nullable ArrayList<SVGLength> mDeltaX;
-    @Nullable ArrayList<SVGLength> mDeltaY;
+    private AlignmentBaseline mAlignmentBaseline;
+    @Nullable private ArrayList<SVGLength> mPositionX;
+    @Nullable private ArrayList<SVGLength> mPositionY;
+    @Nullable private ArrayList<SVGLength> mRotate;
+    @Nullable private ArrayList<SVGLength> mDeltaX;
+    @Nullable private ArrayList<SVGLength> mDeltaY;
     double cachedAdvance = Double.NaN;
 
     public TextView(ReactContext reactContext) {