+++ ios/Pods/Branch/Branch-SDK/Branch-SDK/BNCPreferenceHelper.h
@@ -94,5 +94,6 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory(void);
 
 - (NSMutableString*) sanitizedMutableBaseURL:(NSString*)baseUrl;
 - (void) synchronize;  //  Flushes preference queue to persistence.
++ (void) clearAll;
 
 @end